Fix change password UI
