+++ webapp/src/components/AppBar/UserMenu.jsx
@@ -17,6 +17,7 @@ import { useShowAppDialog } from '../../app/AppDialogServiceProvider';
 import ApiTokenDialog from '../Dialogs/ApiTokenDialog';
 import ChangePasswordDialog from '../Dialogs/ChangePasswordDialog';
 import ContentDialog from '../Dialogs/ContentDialog';
+import showStandardSnackbar from '../../app/showStandardSnackbar';
 
 const ID_MENU_APPBAR = 'menu-appbar';
 const REL = 'noopener noreferrer';
@@ -55,11 +56,20 @@ function UserMenu(props) {
     window.location.reload();
   };
 
-  const onChangePassword = async () => {
+  const onChangePassword = () => {
     showAppDialog(ChangePasswordDialog, {
-      title: 'Change Password',
-      content: 'test',
-      userId: user.id,
+      dialogTitle: 'Change Password',
+      onSubmit: async (onCloseDialog, password) => {
+        try {
+          await usersApi.changePassword(user.id, `"${password}"`);
+          showStandardSnackbar('Password successfully changed.');
+        } catch (e) {
+          showStandardSnackbar(
+            `Could not change password! Reason: ${e.body.message}`
+          );
+        }
+        onCloseDialog();
+      },
     });
   };
 
@@ -92,9 +102,9 @@ function UserMenu(props) {
       <MenuItem onClick={onUserTokenClick}>Get User API Token</MenuItem>
       <MenuItem onClick={onProjectTokenClick}>Get Project API Token</MenuItem>
       <MenuItem onClick={onApiTokenClick}>{t('api_tokens')}</MenuItem>
-      <MenuItem onClick={onChangePassword} disabled={user.has_password}>
-        Change Password
-      </MenuItem>
+      {user?.has_password ? (
+        <MenuItem onClick={onChangePassword}>Change Password</MenuItem>
+      ) : null}
       <MenuItem onClick={onLogoutClick}>{t('logout')}</MenuItem>
       <Divider />
     </div>
@@ -125,7 +135,7 @@ function UserMenu(props) {
         onClose={onClose}
       >
         {isAuthenticated ? privateElements : false}
-        <MenuItem
+        {/* <MenuItem
           className={`${className} menuItem`}
           component="a"
           href={DOCUMENTATION_URL}
@@ -133,7 +143,7 @@ function UserMenu(props) {
           target="_blank"
         >
           {t('documentation')}
-        </MenuItem>
+        </MenuItem> */}
         <MenuItem
           className={`${className} menuItem`}
           component="a"
